+++ conf/init.sql
@@ -27,7 +27,7 @@ CREATE TABLE IF NOT EXISTS device_simulator.generator
     min_value           DOUBLE PRECISION NOT NULL, -- Minimum value for the generated data
     max_value           DOUBLE PRECISION NOT NULL, -- Maximum value for the generated data
     integer             BOOL NOT NULL,
-    interval_seconds    INTEGER NOT NULL,          -- Interval in seconds for data generation
+    interval_seconds    DOUBLE PRECISION NOT NULL, -- Interval in seconds for data generation
     frequency           DOUBLE PRECISION NOT NULL, -- Frequency for wave functions or duty cycle
     initialized_at      TIMESTAMP NOT NULL DEFAULT NOW()
 );
